modules/ai-generation/npc/quest_progression.py
import os
import random
import uuid
import json
import logging
import numpy as np
import tensorflow as tf
from typing import Dict, List, Any, Optional, Tuple

class QuestProgressionSystem:
    """
    Advanced Quest Progression and Player Interaction System
    
    Features:
    - Dynamic quest state management
    - Player behavior analysis
    - Adaptive quest modification
    - Contextual feedback generation
    - Machine learning-driven progression
    """
    
    # Quest States
    QUEST_STATES = [
        'not_started', 'in_progress', 'completed', 
        'failed', 'abandoned', 'on_hold'
    ]
    
    # Player Interaction Types
    INTERACTION_TYPES = [
        'dialogue', 'combat', 'exploration', 
        'crafting', 'trading', 'puzzle_solving'
    ]
    
    def __init__(
        self, 
        models_dir: str = '/opt/mlsp/quest_progression_models',
        cache_dir: str = '/tmp/mlsp_quest_progression_cache'
    ):
        """
        Initialize Quest Progression System
        
        :param models_dir: Directory for storing ML models
        :param cache_dir: Temporary directory for quest progression data
        """
        # Create necessary directories
        os.makedirs(models_dir, exist_ok=True)
        os.makedirs(cache_dir, exist_ok=True)
        
        # Configuration
        self.models_dir = models_dir
        self.cache_dir = cache_dir
        
        # Logging setup
        self.logger = logging.getLogger('QuestProgressionSystem')
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        )
        
        # Load or create ML models
        self._load_models()
    
    def _load_models(self):
        """
        Load or create machine learning models for quest progression
        """
        try:
            # Player Behavior Prediction Model
            behavior_model_path = os.path.join(self.models_dir, 'player_behavior_model.h5')
            if os.path.exists(behavior_model_path):
                self.behavior_model = tf.keras.models.load_model(behavior_model_path)
            else:
                self.behavior_model = self._create_player_behavior_model()
            
            # Quest Progression Prediction Model
            progression_model_path = os.path.join(self.models_dir, 'quest_progression_model.h5')
            if os.path.exists(progression_model_path):
                self.progression_model = tf.keras.models.load_model(progression_model_path)
            else:
                self.progression_model = self._create_quest_progression_model()
        
        except Exception as e:
            self.logger.error(f"Model loading error: {e}")
            raise
    
    def _create_player_behavior_model(self) -> tf.keras.Model:
        """
        Create a neural network for player behavior prediction
        
        :return: Keras player behavior model
        """
        model = tf.keras.Sequential([
            tf.keras.layers.Dense(128, activation='relu', input_shape=(50,)),
            tf.keras.layers.Dropout(0.3),
            tf.keras.layers.Dense(64, activation='relu'),
            tf.keras.layers.Dense(len(self.INTERACTION_TYPES), activation='softmax')
        ])
        
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def _create_quest_progression_model(self) -> tf.keras.Model:
        """
        Create a neural network for quest progression prediction
        
        :return: Keras quest progression model
        """
        model = tf.keras.Sequential([
            tf.keras.layers.Dense(128, activation='relu', input_shape=(60,)),
            tf.keras.layers.Dropout(0.4),
            tf.keras.layers.Dense(64, activation='relu'),
            tf.keras.layers.Dense(len(self.QUEST_STATES), activation='softmax')
        ])
        
        model.compile(
            optimizer='adam',
            loss='categorical_crossentropy',
            metrics=['accuracy']
        )
        
        return model
    
    def track_quest_progress(
        self, 
        quest: Dict[str, Any], 
        player_data: Dict[str, Any]
    ) -> Dict[str, Any]:
        """
        Track and update quest progress
        
        :param quest: Current quest metadata
        :param player_data: Player interaction and progress data
        :return: Updated quest progression metadata
        """
        # Predict player behavior
        player_behavior = self._predict_player_behavior(player_data)
        
        # Update quest objectives
        updated_objectives = self._update_quest_objectives(
            quest['objectives'], 
            player_behavior
        )
        
        # Predict quest progression state
        progression_state = self._predict_quest_progression(
            quest, player_data, player_behavior
        )
        
        # Generate contextual feedback
        feedback = self._generate_quest_feedback(
            quest, 
            updated_objectives, 
            progression_state
        )
        
        # Update quest metadata
        updated_quest = quest.copy()
        updated_quest.update({
            'objectives': updated_objectives,
            'state': progression_state,
            'player_behavior': player_behavior,
            'feedback': feedback
        })
        
        # Save quest progression
        self._save_quest_progression(updated_quest)
        
        return updated_quest
    
    def _predict_player_behavior(
        self, 
        player_data: Dict[str, Any]
    ) -> Dict[str, float]:
        """
        Predict player behavior and interaction type
        
        :param player_data: Player interaction and progress data
        :return: Predicted player behavior probabilities
        """
        # Prepare input features
        input_features = np.concatenate([
            [player_data.get(key, random.random()) for key in [
                'exploration_progress', 'combat_skill', 
                'crafting_level', 'trading_experience'
            ]],
            [random.random() for _ in range(46)]  # Additional randomness
        ])
        
        # Predict behavior
        behavior_probs = self.behavior_model.predict(
            input_features.reshape(1, -1)
        )[0]
        
        return dict(zip(self.INTERACTION_TYPES, behavior_probs))
    
    def _update_quest_objectives(
        self, 
        objectives: List[Dict[str, Any]], 
        player_behavior: Dict[str, float]
    ) -> List[Dict[str, Any]]:
        """
        Update quest objectives based on player behavior
        
        :param objectives: Current quest objectives
        :param player_behavior: Predicted player behavior
        :return: Updated quest objectives
        """
        updated_objectives = []
        
        for objective in objectives:
            if not objective['completed']:
                # Adaptive objective completion probability
                completion_chance = sum(
                    player_behavior.get(interaction_type, 0) 
                    for interaction_type in self.INTERACTION_TYPES
                ) / len(self.INTERACTION_TYPES)
                
                # Probabilistic objective completion
                if random.random() < completion_chance:
                    objective['completed'] = True
            
            updated_objectives.append(objective)
        
        return updated_objectives
    
    def _predict_quest_progression(
        self, 
        quest: Dict[str, Any],
        player_data: Dict[str, Any],
        player_behavior: Dict[str, float]
    ) -> str:
        """
        Predict quest progression state
        
        :param quest: Current quest metadata
        :param player_data: Player interaction and progress data
        :param player_behavior: Predicted player behavior
        :return: Predicted quest progression state
        """
        # Prepare input features
        input_features = np.concatenate([
            [
                quest.get('difficulty_score', random.random()),
                player_data.get('time_spent', random.random()),
                player_data.get('progress_percentage', random.random())
            ],
            list(player_behavior.values()),
            [random.random() for _ in range(50)]  # Additional randomness
        ])
        
        # Predict progression state
        progression_probs = self.progression_model.predict(
            input_features.reshape(1, -1)
        )[0]
        
        # Determine progression state
        progression_state = random.choices(
            self.QUEST_STATES, 
            weights=progression_probs
        )[0]
        
        return progression_state
    
    def _generate_quest_feedback(
        self, 
        quest: Dict[str, Any], 
        objectives: List[Dict[str, Any]],
        progression_state: str
    ) -> Dict[str, Any]:
        """
        Generate contextual quest feedback
        
        :param quest: Current quest metadata
        :param objectives: Updated quest objectives
        :param progression_state: Current quest progression state
        :return: Contextual quest feedback
        """
        feedback = {
            'overall_progress': sum(
                1 for obj in objectives if obj['completed']
            ) / len(objectives) * 100,
            'state_description': self._get_state_description(progression_state),
            'recommendations': self._generate_quest_recommendations(
                quest, objectives, progression_state
            )
        }
        
        return feedback
    
    def _get_state_description(
        self, 
        progression_state: str
    ) -> str:
        """
        Get descriptive text for quest progression state
        
        :param progression_state: Current quest progression state
        :return: State description
        """
        state_descriptions = {
            'not_started': "You haven't begun this quest yet. Are you ready?",
            'in_progress': "You're making progress on this quest. Keep going!",
            'completed': "Congratulations! You've successfully completed the quest.",
            'failed': "Unfortunately, you were unable to complete this quest.",
            'abandoned': "You've decided to abandon this quest. Maybe next time?",
            'on_hold': "This quest is currently on hold. Return when you're ready."
        }
        
        return state_descriptions.get(
            progression_state, 
            "Quest status unknown."
        )
    
    def _generate_quest_recommendations(
        self, 
        quest: Dict[str, Any], 
        objectives: List[Dict[str, Any]],
        progression_state: str
    ) -> List[str]:
        """
        Generate quest-specific recommendations
        
        :param quest: Current quest metadata
        :param objectives: Updated quest objectives
        :param progression_state: Current quest progression state
        :return: List of quest recommendations
        """
        recommendations = []
        
        if progression_state == 'in_progress':
            incomplete_objectives = [
                obj['description'] 
                for obj in objectives 
                if not obj['completed']
            ]
            
            if incomplete_objectives:
                recommendations.append(
                    f"Focus on these remaining objectives: {', '.join(incomplete_objectives)}"
                )
        
        elif progression_state == 'failed':
            recommendations.append(
                f"Review the quest requirements for '{quest['title']}' and try again."
            )
        
        elif progression_state == 'on_hold':
            recommendations.append(
                "Gather more resources or improve your skills before attempting this quest."
            )
        
        return recommendations
    
    def _save_quest_progression(
        self, 
        quest_progression: Dict[str, Any]
    ):
        """
        Save quest progression metadata
        
        :param quest_progression: Updated quest progression metadata
        """
        try:
            progression_path = os.path.join(
                self.cache_dir, 
                f'quest_progression_{quest_progression["id"]}.json'
            )
            
            with open(progression_path, 'w') as f:
                json.dump(quest_progression, f, indent=2)
            
            self.logger.info(f"Quest progression saved: {quest_progression['title']}")
        
        except Exception as e:
            self.logger.error(f"Quest progression save error: {e}")

def main():
    """Example usage of Quest Progression System"""
    # Simulated quest
    quest = {
        'id': str(uuid.uuid4()),
        'title': 'Explore the Forgotten Ruins',
        'difficulty_score': 0.7,
        'objectives': [
            {'description': 'Find the ancient artifact', 'completed': False},
            {'description': 'Defeat the guardian', 'completed': False}
        ]
    }
    
    # Simulated player data
    player_data = {
        'exploration_progress': 0.6,
        'combat_skill': 0.8,
        'time_spent': 1200,  # seconds
        'progress_percentage': 0.4
    }
    
    quest_progression = QuestProgressionSystem()
    
    # Track quest progress
    updated_quest = quest_progression.track_quest_progress(quest, player_data)
    print("\nUpdated Quest Progression:")
    print(json.dumps(updated_quest, indent=2))

if __name__ == '__main__':
    main()
